Field: add more props
- description: An arbitrary string that helps the user use the
    input. Not displayed anywhere by default, but can be used in
    computed props.

Not sure how I feel about this. Maybe we should allow the user
to store an arbitrary "info" object in the component's props?

- focus: Whether or not this input is focused
- focusTime: If focused, the timestamp of when it got focused
- hover,
- hoverTime: same thing as focus, but for hover

The timestamps are useful when one input is focused and another is
hovered. That way, we can tell which happened more recently. If
the hover ends and the focus is still there, we still have info
about the focused element.