Babelify on npm install
This adds a `prepare` script runs `babel-cli` on the src/ dir and
outputs it to the dist/ dir.

This also changes the main file in package.json to target
`dist/Formy/formy.js` instead of the actual source. This is so that when
people install formy as an npm module, they'll have a browser compatible
module to link against in webpack or rollup or whatever.