Build: only transpile with babel
For the built version, don't pack it with webpack (The react-scripts
default). We don't want to package the whole thing bundled with its
large deps as a black box. We want to let the require-er have control
over the exact versions and de-duping the deps.

Like all our other React library repos, just transpile with
babel and let the parent module deal with webpacking.