package.json: move react to peer deps
React is one of the few libraries that can't handle multiple
versions being used at the same time. Thus we need the caller
to be in control of which version; hence the peer dep.

Since 'npm start' and the dev-server still need react,
we'll keep it as a dev dep.

Note: We only ever require 'react' so we don't need to bring
along react-dom. Thought we do need it for the example app.